# go-mux-server/src/api/api_users.go
/*
 * ABM de Usuarios
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package api

import (
	"encoding/json"
	"fmt"
	"net/http"
	"sync"
)

var users []*User
var mutex = &sync.Mutex{}

func UpdateUserByUsername(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	username := getPathParam(w, r)
	var user User
	err := json.NewDecoder(r.Body).Decode(&user)
	if err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}
	mutex.Lock()
	_user, _ := findUserByUsername(*username)
	mutex.Unlock()
	if _user == nil {
		w.WriteHeader(http.StatusNotFound)
		return
	}
	*_user = user
	w.WriteHeader(http.StatusOK)
	fmt.Fprintf(w, toJson(_user))
}

func CreateUser(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	var user User
	err := json.NewDecoder(r.Body).Decode(&user)
	if err != nil || usuario_vacio(&user) {
		http.Error(w, "", http.StatusBadRequest)
		return
	}
	mutex.Lock()
	users = append(users, &user)
	mutex.Unlock()
	w.WriteHeader(http.StatusOK)
}

func DeleteUserByUsername(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	username := getPathParam(w, r)
	mutex.Lock()
	removeByUsername(*username)
	mutex.Unlock()
	w.WriteHeader(http.StatusOK)
}

func GetUserByUsername(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	username := getPathParam(w, r)
	mutex.Lock()
	usuario, _ := findUserByUsername(*username)
	mutex.Unlock()
	if usuario == nil {
		w.WriteHeader(http.StatusNotFound)
		return
	}
	w.WriteHeader(http.StatusOK)
	fmt.Fprintf(w, toJson(usuario))
}

func GetUsers(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusOK)
	if len(users) == 0 {
		fmt.Fprintf(w, "[]")
		return
	}
	fmt.Fprintf(w, generateUserJSON())
}

func usuario_vacio(usuario *User) bool {
	return usuario.Username == "" || usuario.Email == ""
}